requirements.txt
streamlit
yfinance
langchain
langchain-community
langchain-openai
openai
python-dotenv
